# Conflicts:
#	app/src/main/java/org/cxct/sportlottery/db/dao/UserInfoDao.kt
#	app/src/main/java/org/cxct/sportlottery/db/entity/UserInfo.kt
#	app/src/main/java/org/cxct/sportlottery/ui/login/signUp/RegisterActivity.kt
#	app/src/main/java/org/cxct/sportlottery/ui/login/signUp/RegisterViewModel.kt
#	app/src/main/res/values-en/strings.xml
#	app/src/main/res/values-vi/strings.xml
#	app/src/main/res/values/colors.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/org/cxct/sportlottery/network/bet/add/betReceipt/Receipt.kt
@@ -17,5 +17,5 @@ data class Receipt(
     @Json(name = "totalNum")
     val totalNum: Int?,
     @Json(name = "userPlayAmount")
-    val userPlayAmount: Int?
+    val userPlayAmount: Double?
 )
